+++ tests/test_checks.py
@@ -1,11 +1,15 @@
 import os
+import sys
 import unittest
 from unittest.mock import patch
 
 import django
 from django.conf import settings
-from django.core.checks import Warning, run_checks
+from django.core.checks import Error, Warning, run_checks
 from django.test import SimpleTestCase, override_settings
+
+from debug_toolbar import settings as dt_settings
+from debug_toolbar.apps import debug_toolbar_namespace_check
 
 PATH_DOES_NOT_EXIST = os.path.join(settings.BASE_DIR, "tests", "invalid_static")
 
@@ -255,6 +259,82 @@ class ChecksTestCase(SimpleTestCase):
                     "[HKEY_CLASSES_ROOT\\.js]\n"
                     '"Content Type"="application/javascript"',
                     id="debug_toolbar.W007",
+                )
+            ],
+        )
+
+    # @patch("debug_toolbar.apps.debug_toolbar_namespace_check", [])
+    # def test_check_w008_valid(self):
+    #     self.assertEqual(
+    #         run_checks(),
+    #         [
+    #             Error(
+    #                 "debug toolbar  is not a registered namespace",
+    #                 hint="The Django Debug Toolbar is misconfigured, check the documentation for proper configuration",
+    #                 id="debug_toolbar.W008",
+    #             )
+    #         ],
+    #     )
+
+
+def test_check_w008_valid(self):
+    settings.DEBUG = True
+    dt_settings.get_config()["DEBUG_TOOLBAR_CONFIG"] = "test"
+    errors = debug_toolbar_namespace_check(None)
+    self.assertEqual(
+        errors,
+        [
+            Error(
+                "debug toolbar is not a registered namespace",
+                hint="The Django Debug Toolbar is misconfigured, check the documentation for proper configuration",
+                id="debug_toolbar.W008",
+            )
+        ],
+    )
+
+    @patch("debug_toolbar.apps.debug_toolbar_namespace_check", [])
+    def test_debug_toolbar_namespace_check_with_debug_false_and_config_not_test(
+        self, mocked_get_config
+    ):
+        mocked_get_config.return_value = {"DEBUG_TOOLBAR_CONFIG": "not_test"}
+        settings.DEBUG = False
+        errors = debug_toolbar_namespace_check(None)
+        self.assertEqual(len(errors), 0)
+
+    @patch("debug_toolbar.apps.debug_toolbar_namespace_check", [])
+    def test_debug_toolbar_namespace_check_with_debug_true_and_config_test(
+        self, mocked_get_config
+    ):
+        mocked_get_config.return_value = {"DEBUG_TOOLBAR_CONFIG": "test"}
+        settings.DEBUG = True
+        sys.argv = ["manage.py", "test"]
+        errors = debug_toolbar_namespace_check(None)
+        self.assertEqual(len(errors), 0)
+
+    @patch("debug_toolbar.apps.debug_toolbar_namespace_check", [])
+    def test_debug_toolbar_namespace_check_with_debug_true_and_config_not_test(
+        self, mock_get_config
+    ):
+        mock_get_config.return_value = {"DEBUG_TOOLBAR_CONFIG": "not_test"}
+        settings.DEBUG = True
+        sys.argv = ["manage.py", "test"]
+        errors = debug_toolbar_namespace_check(None)
+        self.assertEqual(len(errors), 0)
+
+    @patch("debug_toolbar.apps.debug_toolbar_namespace_check", [])
+    def test_debug_toolbar_namespace_check_with_debug_false_and_config_test(
+        self, mock_get_config
+    ):
+        mock_get_config.return_value = {"DEBUG_TOOLBAR_CONFIG": "test"}
+        settings.DEBUG = False
+        errors = debug_toolbar_namespace_check(None)
+        self.assertEqual(
+            errors,
+            [
+                Error(
+                    "django debug toolbar  is not a registered namespace ",
+                    hint="The Django Debug Toolbar is misconfigured, check the documentation for proper configuration and run the tests.",
+                    id="debug_toolbar.W008",
                 )
             ],
         )